moved menu to a components section
